+++ backend/docong/src/main/java/com/b5f1/docong/api/dto/response/PomoTimeCountResDto.java
@@ -11,5 +11,5 @@ import lombok.ToString;
 @ToString
 public class PomoTimeCountResDto {
     Integer hour;
-    Long cnt;
+    Double cnt;
 }
